Added an abstract layer for the service
